# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the application."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/jobs/listing": {
      "post": {
        "summary": "Create a job listing",
        "tags": ["Jobs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobListingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all job listings",
        "tags": ["Jobs"],
        "responses": {
          "200": {
            "description": "Job fetched successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/jobs/listing/{id}": {
      "get": {
        "summary": "Get a job listing by ID",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Job fetched successfully"
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "patch": {
        "summary": "Update a job listing by ID",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobListingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job updated successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a job listing by ID",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Job deleted successfully"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/jobs/apply/{jobId}": {
      "post": {
        "summary": "Apply for a job",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job ID to apply for"
          }
        ],
        "responses": {
          "201": {
            "description": "Application submitted successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/jobs/apply/{id}": {
      "post": {
        "summary": "Apply for a job",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Job applied successfully"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/jobs/applicants": {
      "get": {
        "summary": "Get all applied Jobs For Job Seeker",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Page number (starts from 1)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Applied application fetched successfully"
          },
          "404": {
            "description": "Applications not found"
          }
        }
      }
    },
    "/jobs/job-types": {
      "get": {
        "summary": "Get all job types available",
        "tags": ["Jobs"],
        "responses": {
          "200": {
            "description": "Job Types fetched successfully"
          },
          "404": {
            "description": "Job Types not found"
          }
        }
      }
    },
    "/jobs/payment-types": {
      "get": {
        "summary": "Get all payment types available",
        "tags": ["Jobs"],
        "responses": {
          "200": {
            "description": "Payment Types fetched successfully"
          },
          "404": {
            "description": "Payment Types not found"
          }
        }
      }
    },
    "/jobs/job-with-applicants": {
      "get": {
        "summary": "Get all jobs with applicants for Job Provider",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Page number (starts from 1)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs with applican fetched successfully"
          },
          "404": {
            "description": "Jobs not found"
          }
        }
      }
    },
    "/jobs/job-applications": {
      "get": {
        "summary": "Get all jobs with applicants details for Job Provider",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Page number (starts from 1)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs with applican fetched successfully"
          },
          "404": {
            "description": "Jobs not found"
          }
        }
      }
    },
...
    "/jobs/listing/shifts": {
      "get": {
        "summary": "Get user job listing shifts data",
        "tags": ["Jobs"],
        "responses": {
          "200": {
            "description": "User job listing shifts data fetched successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateJobListingDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Software Engineer"
          },
          "description": {
            "type": "string",
            "example": "Job description here"
          },
          "address": {
            "type": "string",
            "example": "123 Main Street"
          },
          "zipCode": {
            "type": "string",
            "example": "12345"
          },
          "type": {
            "type": "string",
            "enum": ["site_worker", "remote_worker"],
            "example": "site_worker"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["construction", "full-time", "urgent"]
          },
          "country": {
            "type": "string",
            "example": "United States"
          },
          "state": {
            "type": "string",
            "example": "California"
          },
          "city": {
            "type": "string",
            "example": "Los Angeles"
          },
          "companyName": {
            "type": "string",
            "example": "BuildCorp Ltd."
          },
          "minimumRequirements": {
            "type": "integer",
            "example": 5
          },
          "maximumRequirements": {
            "type": "integer",
            "example": 50
          },
          "paymentType": {
            "type": "string",
            "enum": ["contractual", "hourly", "monthly"],
            "example": "contractual"
          },
          "payment": {
            "type": "integer",
            "example": 2000
          },
          "shiftStartsAt": {
            "type": "integer",
            "example": 12
          },
          "shiftEndsAt": {
            "type": "integer",
            "example": 6
          }
        }
      },
      "UpdateJobListingDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Updated Job Title"
          },
          "description": {
            "type": "string",
            "example": "Updated job description"
          }
        }
      }
    }
  }
}